integration/apex_code/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "apex-code"
version = "1.0.0"
description = "APEX Code Harness - AI model evaluation framework for software engineering tasks"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "APEX Code Contributors"}
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "docker>=6.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
apx = "apex_code.cli.main:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.setuptools.packages.find]
where = ["."]
include = ["apex_code*"]
